Set the CFDs to only work on baseline subtracted traces.
I have also updated the WaveformAnalyzer and Trace classes to handle things a litlte differently. This has increased the efficiency of the software since we're only calculating the baseline subtraction on the trace once, instead of every time somebody needs it. 

Former-commit-id: 62d0cbcfa3c55d2bfb0b6655dcfc0a4f45a4a84a

+++ Scan/Resources/include/TraditionalCfd.hpp
@@ -16,7 +16,7 @@ public:
     /// Default destructor
     ~TraditionalCfd() {};
 
-    double CalculatePhase(const std::vector<unsigned int> &data,
+    double CalculatePhase(const std::vector<double> &data,
                           const std::pair<double, double> &pars,
                           const std::pair<unsigned int, double> &max,
                           const std::pair<double, double> baseline);
